requirements.txt
Flask
pymongo
python-dotenv
razorpay
requests
Pillow
